Split functions into multiple files, added GitHub calls
