+++ lua/tickets/utils.lua
@@ -1,0 +1,10 @@
+M = {}
+
+function M.expand_path(path)
+    if path:sub(1, 1) == "~" then
+        return os.getenv("HOME") .. path:sub(2)
+    end
+    return path
+end
+
+return M
